refactor: consolidate partition tests into ZonemapFragmentPrunerTest
Move all partition detection and PartitionInfo tests from
ZonemapPartitionDetectorTest into ZonemapFragmentPrunerTest and
delete the now-empty test file.
